Split up TabBarItem class
