+++ AuroraEditor/Base/TabBar/UI/TabBarItem/TabBarItem.swift
@@ -7,32 +7,31 @@
 
 import SwiftUI
 
-// TODO: Disable the rule because this view is fairly complicated and I have already modularize some parts.
-// swiftlint:disable:next type_body_length
 struct TabBarItem: View {
-    @Environment(\.colorScheme)
-    private var colorScheme
-
-    @Environment(\.controlActiveState)
-    private var activeState
 
     @Environment(\.isFullscreen)
     private var isFullscreen
 
+    @Environment(\.controlActiveState)
+    var activeState
+
+    @Environment(\.colorScheme)
+    var colorScheme
+
     @StateObject
-    private var prefs: AppPreferencesModel = .shared
+    var prefs: AppPreferencesModel = .shared
 
     @State
-    private var isHovering: Bool = false
+    var isHovering: Bool = false
 
     @State
-    private var isHoveringClose: Bool = false
+    var isHoveringClose: Bool = false
 
     @State
-    private var isPressingClose: Bool = false
+    var isPressingClose: Bool = false
 
     @State
-    private var isAppeared: Bool = false
+    var isAppeared: Bool = false
 
     @Binding
     private var expectedWidth: CGFloat
@@ -40,17 +39,17 @@ struct TabBarItem: View {
     @ObservedObject
     var workspace: WorkspaceDocument
 
-    private var item: TabBarItemRepresentable
+    var item: TabBarItemRepresentable
 
     private var windowController: NSWindowController
 
-    private var isTemporary: Bool
+    var isTemporary: Bool
 
     var isActive: Bool {
         item.tabID == workspace.selectionState.selectedId
     }
 
-    private func switchAction() {
+    func switchAction() {
         // Only set the `selectedId` when they are not equal to avoid performance issue for now.
         if workspace.selectionState.selectedId != item.tabID {
             workspace.selectionState.selectedId = item.tabID
@@ -93,118 +92,7 @@ struct TabBarItem: View {
                 .opacity(isActive && prefs.preferences.general.tabBarStyle == .xcode ? 0.0 : 1.0)
                 .padding(.top, isActive && prefs.preferences.general.tabBarStyle == .native ? 1.22 : 0)
             // Tab content (icon and text).
-            HStack(alignment: .center, spacing: 5) {
-                item.icon
-                    .resizable()
-                    .aspectRatio(contentMode: .fit)
-                    .foregroundColor(
-                        prefs.preferences.general.fileIconStyle == .color && activeState != .inactive
-                        ? item.iconColor
-                        : .secondary
-                    )
-                    .frame(width: 12, height: 12)
-                Text(item.title)
-                    .font(
-                        isTemporary
-                        ? .system(size: 11.0).italic()
-                        : .system(size: 11.0)
-                    )
-                    .lineLimit(1)
-            }
-            .frame(
-                // To horizontally max-out the given width area ONLY in native tab bar style.
-                maxWidth: prefs.preferences.general.tabBarStyle == .native ? .infinity : nil,
-                // To max-out the parent (tab bar) area.
-                maxHeight: .infinity
-            )
-            .padding(.horizontal, prefs.preferences.general.tabBarStyle == .native ? 28 : 23)
-            .overlay {
-                ZStack {
-                    if isActive {
-                        // Close Tab Shortcut:
-                        // Using an invisible button to contain the keyboard shortcut is simply
-                        // because the keyboard shortcut has an unexpected bug when working with
-                        // custom buttonStyle. This is an workaround and it works as expected.
-                        Button(
-                            action: closeAction,
-                            label: { EmptyView() }
-                        )
-                        .frame(width: 0, height: 0)
-                        .padding(0)
-                        .opacity(0)
-                        .keyboardShortcut("w", modifiers: [.command])
-                    }
-                    // Switch Tab Shortcut:
-                    // Using an invisible button to contain the keyboard shortcut is simply
-                    // because the keyboard shortcut has an unexpected bug when working with
-                    // custom buttonStyle. This is an workaround and it works as expected.
-                    Button(
-                        action: switchAction,
-                        label: { EmptyView() }
-                    )
-                    .frame(width: 0, height: 0)
-                    .padding(0)
-                    .opacity(0)
-                    .keyboardShortcut(
-                        workspace.getTabKeyEquivalent(item: item),
-                        modifiers: [.command]
-                    )
-                    .background(.blue)
-                    // Close button.
-                    Button(action: closeAction) {
-                        if prefs.preferences.general.tabBarStyle == .xcode {
-                            Image(systemName: "xmark")
-                                .font(.system(size: 11.2, weight: .regular, design: .rounded))
-                                .frame(width: 16, height: 16)
-                                .foregroundColor(
-                                    isActive
-                                    ? (
-                                        colorScheme == .dark
-                                        ? .primary
-                                        : Color(nsColor: .controlAccentColor)
-                                    )
-                                    : .secondary.opacity(0.80)
-                                )
-                        } else {
-                            Image(systemName: "xmark")
-                                .font(.system(size: 9.5, weight: .medium, design: .rounded))
-                                .frame(width: 16, height: 16)
-                        }
-                    }
-                    .buttonStyle(.borderless)
-                    .foregroundColor(isPressingClose ? .primary : .secondary)
-                    .background(
-                        colorScheme == .dark
-                        ? Color(nsColor: .white)
-                            .opacity(isPressingClose ? 0.32 : isHoveringClose ? 0.18 : 0)
-                        : (
-                            prefs.preferences.general.tabBarStyle == .xcode
-                            ? Color(nsColor: isActive ? .controlAccentColor : .black)
-                                .opacity(
-                                    isPressingClose
-                                    ? 0.25
-                                    : (isHoveringClose ? (isActive ? 0.10 : 0.06) : 0)
-                                )
-                            : Color(nsColor: .black)
-                                .opacity(isPressingClose ? 0.29 : (isHoveringClose ? 0.11 : 0))
-                        )
-                    )
-                    .cornerRadius(2)
-                    .accessibilityLabel(Text("Close"))
-                    .onHover { hover in
-                        isHoveringClose = hover
-                    }
-                    .pressAction {
-                        isPressingClose = true
-                    } onRelease: {
-                        isPressingClose = false
-                    }
-                    .opacity(isHovering ? 1 : 0)
-                    .animation(.easeInOut(duration: 0.08), value: isHovering)
-                    .padding(.leading, prefs.preferences.general.tabBarStyle == .xcode ? 3.5 : 4)
-                }
-                .frame(maxWidth: .infinity, alignment: .leading)
-            }
+            iconTextView
             .opacity(
                 // Inactive states for tab bar item content.
                 activeState != .inactive
@@ -342,15 +230,3 @@ struct TabBarItem: View {
     }
 }
 // swiftlint:enable type_body_length
-
-fileprivate extension WorkspaceDocument {
-    func getTabKeyEquivalent(item: TabBarItemRepresentable) -> KeyEquivalent {
-        for counter in 0..<9 where self.selectionState.openFileItems.count > counter &&
-        self.selectionState.openFileItems[counter].tabID == item.tabID {
-            return KeyEquivalent.init(
-                Character.init("\(counter + 1)")
-            )
-        }
-        return "0"
-    }
-}
